Modified Camera UBO and added Gaussian buffers
- Refactored Camera class and added camera position.
- Added VmaAllocationCreateFlags parameter to buffer creation functions in ResourceAllocator.
- Added VmaAllocationInfo to Buffer and AccelerationStructure structs in ResourceAllocator.
- Modified uniform buffers and push constants for Gaussian pipeline.

+++ Engine/include/engine_renderer.h
@@ -38,7 +38,7 @@ namespace core {
 
 		// Descriptor buffers.
 		Buffer m_cameraDescBuffer;
-		uint32_t m_cameraDescBufferAlignment;
+		size_t m_cameraDescBufferAlignment;
 
 		// Instances for all the renderers.
 		std::unique_ptr<StandardRenderer> m_standardRenderer;
